refactor: Change file name UsageStatus to UsageHistory and UsageInfo to UsageHistoryInfo
