+++ Views/nav_Bar.ejs
@@ -63,14 +63,14 @@
                   <li class="nav-item" id="resizemenu">
                     <div class="dropdown"><a class="nav-link js-scroll-trigger dropdown-toggle" href="#" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" style="background-color:transparent;">My Warehouse </a>
                       <div class="dropdown-menu" aria-labelledby="dropdownMenuButton" id="accountMenu">
-                        <a class="dropdown-item" href="/Buyer/UsageStatus" id="account_ctrl">Usage Status</a>
+                        <a class="dropdown-item" href="/Buyer/UsageHistory" id="account_ctrl">Usage History</a>
                         <a class="dropdown-item" href="/Buyer/RequestStatus">Request Status</a>
                       </div>
                     </div>
                   </li>
                 </div>
                 <div class="x990">
-                  <li class="nav-item"><a class="nav-link js-scroll-trigger" href="/Buyer/UsageStatus">Usage Status</a></li>
+                  <li class="nav-item"><a class="nav-link js-scroll-trigger" href="/Buyer/UsageHistory">Usage History</a></li>
                   <li class="nav-item"><a class="nav-link js-scroll-trigger" href="/Buyer/RequestStatus">Request Status</a></li>
                 </div>
               <% } %>
